Refactor to game+screen
