+++ space-invaders/core/src/com/samsung/business/spaceinvaders/MainMenuScreen.java
@@ -1,0 +1,44 @@
+package com.samsung.business.spaceinvaders;
+
+import com.badlogic.gdx.Screen;
+
+public class MainMenuScreen implements Screen {
+    public MainMenuScreen(SpaceInvaders spaceInvaders) {
+
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
